\name{nontargetData-package}

\alias{nontargetData-package}

\alias{nontargetData}

\docType{package}

\title{
Quantized simulation data of isotope pattern centroids
}

\description{
Data sets for isotope pattern grouping of LC-HRMS peaks with package nontarget. 
Based on a vast set of PubChem molecular formulas, quantized (a) m/z, (b) m/z differences, (c) intensity ratios and (d) marker centroids
of simulated centroid pairs are listed for different instrument resolutions.
}

\details{
\tabular{ll}{
Package: \tab nontargetData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-22\cr
License: \tab GPL-3\cr
}
}

\author{
	Martin Loos, Francesco Corona
	Maintainer: Martin Loos <Martin.Loos@eawag.ch>
}

\keyword{ package }
